Update CreateDefectInAzureDevops.js
Changed the timeout in getDefectDetailsByIdWithRetry() as Pulse rules naturally time out at 60 seconds and after 4 iterations of the current timeout structure, the rule would cause a PULERR001.  Fixed it to resolve the timeout within the scope of the Pulse rule execution period.

KNOWN ISSUES: There is an associated SWAT with this issue (SWAT-2886) as the webhook event for creating a defect is fired when the "New" button is clicked rather then when the form is submitted, thus if a user does not submit the new defect within 60 seconds the integration will fail.